% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_companion_F.R
\name{build_companion_F}
\alias{build_companion_F}
\title{Build the companion matrix of factors}
\usage{
build_companion_F(F_matrix, p)
}
\arguments{
\item{F_matrix}{A matrix of factors}

\item{p}{An integer. Autoregressive order of the state equation}
}
\value{
A block matrix with the first p lags of the factors in the first row
}
\description{
Build the companion matrix of factors
}
