% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_companion_G.R
\name{build_companion_G}
\alias{build_companion_G}
\title{Build the companion matrix of the covariance matrix of the innovations in the state equation}
\usage{
build_companion_G(Gamma, p)
}
\arguments{
\item{Gamma}{A matrix. Covariance matrix of the innovations in the state equation}

\item{p}{An integer. Autoregressive order of the state equation}
}
\value{
A bolck matrix with the covariance matrix in the first row
}
\description{
Build the companion matrix of the covariance matrix of the innovations in the state equation
}
