% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_companion_Lambda.R
\name{build_companion_Lambda}
\alias{build_companion_Lambda}
\title{Build the companion matrix of the loadings}
\usage{
build_companion_Lambda(Lambda, p)
}
\arguments{
\item{Lambda}{A matrix of loadings}

\item{p}{An integer. Autoregressive order of the state equation}
}
\value{
A bolck matrix with the loadings in the first row
}
\description{
Build the companion matrix of the loadings
}
