% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_to_fit.R
\name{dfm_to_fit}
\alias{dfm_to_fit}
\title{Convert dynfactoR output to emDFM format}
\usage{
dfm_to_fit(dfm_res)
}
\arguments{
\item{dfm_res}{The output of \code{dynfactoR::dfm()}}
}
\value{
A list compatible with \code{emDFM} functions
}
\description{
This function is only for initial development and is likely to
get deprecated in future releases.
}
\examples{
# dfm_res <- dfm(d)
# dfm_fit <- dfm_to_fit(dfm_res)
# plot_forecast(dfm_fit)

}
\seealso{
\code{dynfactoR} package: https://github.com/guilbran/dynfactoR
}
