% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_clean_data.R
\name{download_clean_data}
\alias{download_clean_data}
\title{Download data from Eurostat and Google and clean them}
\usage{
download_clean_data(geo_code)
}
\arguments{
\item{geo_code}{A string containing the country code, for example "IT" for Italy or "FR" for France}
}
\value{
A list of two tibbles, the first with the daily GMI,
the second with an aggregated monthly mobility index
}
\description{
\code{download_clean_data} returns two tibbles containing monthly time series from
Eurostat and the Google Mobility Index (GMI) for a chosen country. Data are first
downloaded from the official APIs and then merged and cleaned so to make all
the series stationary. The first tibble contains the high-frequency GMI, while the
second tibble reports the indexes aggregated to have monthly frequency.
}
\examples{
#country_code <- "IT"
#data <- download_clean_data(country_code)
}
