% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{em_algorithm}
\alias{em_algorithm}
\title{Estimate the Dynamic Factor Model using the Expectation Maximization algorithm}
\usage{
em_algorithm(
  d,
  r = 4,
  p = 1,
  thresh = 0.01,
  thresh_imputation = 0.01,
  show_plots = FALSE,
  max_iter = 30
)
}
\arguments{
\item{d}{A tibble of time series, with dates in the first column.}

\item{r}{An integer. Number of factors to estimate.}

\item{p}{An integer. Autoregressive order of the state equation.}

\item{thresh}{A double. Threshold for convergence of the EM algorithm.}

\item{thresh_imputation}{A double. Threshold for the imputation of missing values (for algorithm initialization).}

\item{show_plots}{Logical. Whether to show or not the plot of the factors at each iteration of the EM algorithm.
Default is FALSE.}

\item{max_iter}{An integer. Maximum number of iterations in the EM algorithm. Default is 30.}
}
\value{
A list of parameters:
\item{\code{factors}}{A matrix of estimated factors}
\item{\code{P_list}}{List of estimated variance of the forecasts in Kalman filter update}
\item{\code{F_prediction_list}}{List of predicted factors in the Kalman filter update}
\item{\code{P_prediction_list}}{List of predicted variance of the forecasts in Kalman filter update}
\item{\code{parameters}}{A list containing the parameters of the model.
\code{Lambda} is the matrix of loadings; \code{A} is the list of the matrices of the autoregressive parameters at each lag;
\code{p} is the autoregressive order of the state equation; \code{Gamma_eta} is the covariance matrix of the innovations in the state equation}
}
\description{
Estimate the Dynamic Factor Model using the Expectation Maximization algorithm
}
\examples{
# d <- download_clean_data()
# em_fit <- em_algorithm(d)
# em_fit$factors

}
