% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_values.R
\name{fitted_values}
\alias{fitted_values}
\title{Get the fitted values of the Dynamic Factor Model}
\usage{
fitted_values(obj_fit, d)
}
\arguments{
\item{obj_fit}{A list of estimated parameters of the Dynamic Factor Model, as
obtained from \code{pca_estimator()}, \code{kalman_filter()}, \code{kalman_smoother()}, \code{em_algorithm()}}

\item{d}{A tibble of time series, with dates in the first column}
}
\value{
A matrix of fitted values of the DFM
}
\description{
Get the fitted values of the Dynamic Factor Model
}
\examples{
# d <- download_clean_data("IT")
# em_fit <- em_algorithm(d)
# rmse_fit(em_fit, d)

}
\seealso{
\code{pca_estimator}, \code{initialize_filter}, \code{kalman_filter}, \code{kalman_smoother}, \code{em_algorithm}
}
