% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Get forecasts form the estimated DFM}
\usage{
forecast(obj_fit, horizon = 1)
}
\arguments{
\item{obj_fit}{A list of estimated parameters of the Dynamic Factor Model, as
obtained from \code{pca_estimator()}, \code{kalman_filter()}, \code{kalman_smoother()}, \code{em_algorithm()}}

\item{horizon}{An integer. Number of steps-ahead of the forecast. Default is 1.}
}
\value{
A vector of \code{horizon}-steps-ahead forecasts for all the variables in the original dataset
}
\description{
Get forecasts form the estimated DFM
}
\examples{
# d <- download_clean_data("IT")
# em_fit <- em_algorithm(d)
# forecast(em_fit, horizon = 1)

}
