% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_evaluation_aux.R
\name{forecast_evaluation_aux}
\alias{forecast_evaluation_aux}
\title{Wrapper for forecast evaluation}
\usage{
forecast_evaluation_aux(type, evaluation_input)
}
\arguments{
\item{type}{String to choose the algorithm. Either "PCA", "KF", "KS", "EM"}

\item{evaluation_input}{A list}
}
\value{
See \code{forecast_evaluation}
}
\description{
Wrapper for forecast evaluation
}
\examples{
# d <- download_clean_data("IT")
# evaluation_input <- list(d = d,
#                          horizon = 1,
#                          window_size = 300
#                          r = 4,
#                          p = 1,
#                          thresh_imputation = 0.2,
#                          thresh = 0.01                 # L1 convergence of parameters in EM
#                          )
#
# forecast_eval <- forecast_evaluation_aux("EM", evaluation_input)
# plot_forecast_evaluation(forecast_eval)

}
\seealso{
\code{forecast_evaluation}
}
