% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_experiment.R
\name{forecast_experiment}
\alias{forecast_experiment}
\title{Perform forecast experiment and compare different algorithms}
\usage{
forecast_experiment(
  evaluation_input,
  plot_variable = "indpro_SCA",
  ggplot_theme = NULL
)
}
\arguments{
\item{evaluation_input}{A list}

\item{plot_variable}{String defining the variable to plot}

\item{ggplot_theme}{ggplot2 theme (optional)}
}
\value{
A list
\item{\code{rmsfe}}{A tibble containing the RMSFE for each variable and for each estimation algorithm.}
\item{\code{plot_forecast}}{Plot of the forecasted vs. observed values for all the estimation algorithm for the selected variable.}
\item{\code{plot_overall_rmsfe}}{Plot of the average RMSFE for each estimation algorithm.}
\item{\code{plot_disagg_rmsfe}}{Plot of the RMSFE for each estimation algorithm for a subset of the variables in the dataset}
}
\description{
Perform forecast experiment and compare different algorithms
}
\examples{
# d <- download_clean_data("IT")
# evaluation_input <- list(d = d,
#                          horizon = 1,
#                          window_size = 300
#                          r = 4,
#                          p = 1,
#                          thresh_imputation = 0.2,
#                          thresh = 0.01                 # L1 convergence of parameters in EM
#                          )
# experiment_result <- forecast_experiment(evaluation_input, plot_variable = "indpro_SCA")

}
