% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_pipeline.R
\name{forecast_pipeline}
\alias{forecast_pipeline}
\title{Forecast using the Dynamic Factor Model}
\usage{
forecast_pipeline(dataset, forecast_input)
}
\arguments{
\item{dataset}{A tibble of time series, with dates in the first column}

\item{forecast_input}{A list}
}
\value{
A plot of the last observed values and the forecasts
}
\description{
Forecast using the Dynamic Factor Model
}
\examples{
# d <- download_clean_data("IT")
# eurostat_indpro_NSA <- get_eurostat("sts_inpr_m",
# filters = list(geo = input$geo_code,
#               s_adj = "NSA",
#               unit = "I15",
#               nace_r2 = "C")) \%>\%
#  dplyr::select(date = time, indpro_NSA = values)
# seasonal_attributes_indpro <- get_seasonal_attributes_indpro(eurostat_indpro_NSA)
# forecast_input <- list(estimator = "EM",
#                        variable = "indpro_SCA",
#                        horizon = 6,
#                        r = 4,
#                        p = 1,
#                        thresh = 0.01,
#                        thresh_imputation = 0.02,
#                        seasonal_attributes_indpro = seasonal_attributes_indpro
# forecast_plot <- forecast_pipeline(dataset, forecast_input)


}
