% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seasonal_attributes_indpro.R
\name{get_seasonal_attributes_indpro}
\alias{get_seasonal_attributes_indpro}
\title{Get seasonal attributes for industrial production}
\usage{
get_seasonal_attributes_indpro(indpro_NSA_tibble)
}
\arguments{
\item{indpro_NSA_tibble}{A tibble containing a column \code{date} and a column \code{indpro_NSA}}
}
\value{
A tibble containing the dates and the trend and seasonal components of industrial production
#' @importFrom lubridate \%m+\% NULL
}
\description{
Get the trend, seasonal and residuals components from the tibble of
industrial production obtained via eurostat::get_eurostat().
}
\details{
Please notice that this function is only for initial development and is
likely to be deprecated in future releases
}
