% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_sample_evaluation.R
\name{in_sample_evaluation}
\alias{in_sample_evaluation}
\title{Perform in-sample evaluation of the DFM estimated via different algorithms}
\usage{
in_sample_evaluation(
  d_scaled,
  ggplot_theme = NULL,
  r = 4,
  p = 1,
  thresh_imputation = 0.01,
  thresh = 0.01
)
}
\arguments{
\item{d_scaled}{A tibble of time series, with dates in the first column}

\item{ggplot_theme}{ggplot2 theme for plots (optional)}

\item{r}{An integer. Number of factors in the DFM}

\item{p}{An integer. Autoregressive order of the state equation}

\item{thresh_imputation}{A double. Threshold for the imputation of missing values.}

\item{thresh}{A double. Threshold for convergence of the EM algorithm.}
}
\value{
A list:
\item{\code{rmse_in_sample}}{Tibble of in-sample RMSEs for each algorithm and for each variable in the dataset}
\item{\code{plot_overall_rmse}}{Plot of the average RMSE for each algorithm}
\item{\code{plot_disagg_rmse}}{Plot of the RMSEs for each algorithm on a subset of the variables in the dataset}
}
\description{
\code{in_sample_evaluation} takes a tibble of time series and compares the in-sample fit of the Dynamic Factor Model
using the four different algorithm in \code{pca_estimator}, \code{kalman_filter}, \code{kalman_smoother}, \code{em_algorithm}.
For each estimation method, it reports the values of the RMSE for each variable and plots them.
}
\examples{
# d <- download_clean_data("IT")
# in_sample_evaluation(d)

}
\seealso{
\code{pca_estimator}, \code{initialize_filter}, \code{kalman_filter}, \code{kalman_smoother}, \code{em_algorithm}
}
