% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_filter.R
\name{initialize_filter}
\alias{initialize_filter}
\title{Initialize a list of parameters for Kalman filter estimation}
\usage{
initialize_filter(pc_fit, restrict_Gamma_eta = FALSE)
}
\arguments{
\item{pc_fit}{A list. The output of \code{pca_estimator}.}

\item{restrict_Gamma_eta}{Logical. If TRUE, the covariance matrix of the innovations in the state
equation is constrained to be the identity matrix, i.e. the innovations are assumed to be
uncorrelated with unit variance. Default is FALSE.}
}
\value{
A list of parameters for Kalman filter initialization:
\item{\code{Lambda}}{Matrix of loadings estimated by \code{pca_estimator}.}
\item{\code{A}}{List of matrices of the autoregressive parameters in the state equation, estimated via \code{pca_estimator}.}
\item{\code{Gamma_csi}}{A matrix containing the diagonal of the estimated covariance matrix of the measurement errors.}
\item{\code{Gamma_eta}}{A matrix containing the estimated covariance matrix of the factors' innovations.
If \code{restrict_Gamma_eta} is TRUE, then it is an identity matrix.}
\item{\code{F0}}{A matrix of factors, as estimated by \code{pca_estimator}.}
\item{\code{p}}{An integer. Autoregressive order of the state equation.}
}
\description{
Initialize a list of parameters for Kalman filter estimation
}
\examples{
# pc_fit <- pca_estimator(d)
# param_list <- initializa_filter(pc_fit)
# kf_fit <- kalman_filter(d, param_list)

}
\seealso{
\code{pca_estimator}
}
