% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{kalman_filter}
\alias{kalman_filter}
\title{Kalman filter estimation of the Dynamic Factor Model}
\usage{
kalman_filter(d, parameters)
}
\arguments{
\item{d}{A tibble of time series, with dates in the first column}

\item{parameters}{A list of parameters for filter initialization, obtained with \code{initialize_filter()}}
}
\value{
A list of parameters:
\item{\code{factors}}{A matrix of estimated factors}
\item{\code{P_list}}{Estimated variance of the forecasts in Kalman filter update}
\item{\code{F_prediction_list}}{Predicted factors in the Kalman filter update}
\item{\code{P_prediction_list}}{Predicted variance of the forecasts in Kalman filter update}
\item{\code{parameters}}{A list containing the parameters of the model.
\code{Lambda} is the matrix of loadings; \code{A} is the list of the matrices of the autoregressive parameters at each lag;
\code{p} is the autoregressive order of the state equation; \code{Gamma_eta} is the covariance matrix of the innovations in the state equation}
}
\description{
Kalman filter estimation of the Dynamic Factor Model
}
\examples{
# pc_fit <- pca_estimator(d)
# param_list <- initializa_filter(pc_fit)
# kf_fit <- kalman_filter(d, param_list)

}
