% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_smoother.R
\name{kalman_smoother}
\alias{kalman_smoother}
\title{Kalman smoother estimation of the Dynamic Factor Model}
\usage{
kalman_smoother(kalman_filter_fit)
}
\arguments{
\item{kalman_filter_fit}{A list obtained from \code{kalman_filter()}}
}
\value{
A list \code{kalman_filter_fit} with factors updated using the Kalman smoother algorithm
}
\description{
Kalman smoother estimation of the Dynamic Factor Model
}
\examples{
# pc_fit <- pca_estimator(d)
# param_list <- initializa_filter(pc_fit)
# kf_fit <- kalman_filter(d, param_list)
# ks_fit <- kalman_smoother(kf_fit)

}
