% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_to_missing.R
\name{outliers_to_missing}
\alias{outliers_to_missing}
\title{Substitute outliers with missing values}
\usage{
outliers_to_missing(d)
}
\arguments{
\item{d}{A tibble of time series containing a column \code{date}}
}
\value{
A tibble of the same dimension, where outliers has been substituted with NA
}
\description{
Converts outliers in a tibble to NA. Outliers are defined as observations
that deviate from the column mean by more than 6 inter-quartile ranges.
}
\examples{
# outliers_to_missing(d)

}
