% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_estimator.R
\name{pca_estimator}
\alias{pca_estimator}
\title{Estimation of the DFM using PCA}
\usage{
pca_estimator(d, r = 4, p = 1)
}
\arguments{
\item{d}{A tibble of time series data, with dates in the first column}

\item{r}{The number of factors to extract}

\item{p}{The autoregressive order of the state equation}
}
\value{
A list of length 4 containing the estimated DFM
\item{\code{factors}}{A matrix containing the estimated factors}
\item{\code{parameters}}{A list containing the parameters of the model.
\code{Lambda} is the matrix of loadings; \code{A} is the list of the matrices of the autoregressive parameters at each lag;
\code{p} is the autoregressive order of the state equation; \code{Gamma_eta} is the covariance matrix of the innovations in the state equation}
\item{\code{data_matrix}}{The matrix containing the time series (without dates)}
\item{\code{chi}}{Common component of the time series, obtained as the reconstruction using the first \code{r} principal components}
}
\description{
\code{pca_estimator} takes as input a tibble of time series with the dates
in the first column. If there are missing values, these are omitted (even if it is possible to impute them in
advance using \code{scale_impute}).
Then the first \code{r} principal components are extracted and a VAR of order \code{p} is fitted on them
to estimate the autoregressive coefficients of the state equation and the covariance matrix of
factors innovations.
}
\examples{
# pc_fit <- pca_estimator(d)
# pc_fit$factors

}
\seealso{
\code{scale_impute}
}
