% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast_evaluation.R
\name{plot_forecast_evaluation}
\alias{plot_forecast_evaluation}
\title{Plot the results of a forecast evaluation of a DFM}
\usage{
plot_forecast_evaluation(
  forecast_eval,
  variable = "indpro_SCA",
  title = NULL,
  annotate = TRUE
)
}
\arguments{
\item{forecast_eval}{A list, output of \code{forecast_evaluation}.}

\item{variable}{A string defining the target variable.}

\item{title}{A string with the title to assign to the plot (optional).}

\item{annotate}{A string with annotations for the plot (optional)}
}
\value{
A plot comparing forecasted values and ground truth.
}
\description{
Plot the results of a forecast evaluation of a DFM
}
\examples{
# d <- download_clean_data("IT")
# eval_obj <- forecast_evaluation(d, horizon = 1, window_size = 300, type = "EM")
# plot_forecast_evaluation(eval_obj, variable = "indpro_SCA",
#                          title = "Forecast VS ground truth, Italian industrial production")


}
