% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_impute.R
\name{scale_impute}
\alias{scale_impute}
\title{Scale data and impute missing values}
\usage{
scale_impute(d, r = 4, thresh = 0.01)
}
\arguments{
\item{d}{A tibble of time series, having dates as the first column}

\item{r}{An integer. Number of principal components to use in imputation}

\item{thresh}{A double defining the threshold for convergence of the algorithm}
}
\value{
A tibble where missing values in \code{d} have been imputed
}
\description{
Imputation of missing values follows the procedure described in Mc Cracken and Ng (2015).
It is based on the EM algorithm in Stock and Watson (2002):
\itemize{
\item Scale data
\item Initialize missing values to zero (column mean)
\item Extract the first \code{r} principal components
\item Compute the reconstruction of the dataset based on the \code{r} principal components
\item Impute the missing values with their reconstructed values
\item Repeat PCA and imputation until convergence of the principal components
}
}
\examples{
# scale_impute(d)

}
