% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_tibble.R
\name{scale_tibble}
\alias{scale_tibble}
\title{Scale data in a tibble, excluding the first column}
\usage{
scale_tibble(d_tibble)
}
\arguments{
\item{d_tibble}{A tibble of time series having dates as the first column}
}
\value{
A list:
\item{\code{scaled_tibble}}{Tibble of the scaled data}
\item{\code{scale_attributes}}{A list containing the vector of column means and standard deviations of the
original tibble.}
}
\description{
Scale data in a tibble, excluding the first column
}
\examples{
#scale_tibble(d)

}
